% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_lattice.R
\name{binomialLattice}
\alias{binomialLattice}
\title{n = Anzahl Perioden
S0 = Heutige Aktienpreis
K = Strike
sigma = Volatilität
T = Periodizität: quartalsweise : T= 4, halbjährlich T = 2
optionType = Call/Put
exerciseType = Eu/Ami für Ami --> "American"
u = Preissprung nach oben
d = Preissprung nach unten
r = risikofreier Zins
v = erwarte Wachstumsrate der log-preise
Zeitschritte anwenden
Berechne Preissprünge oben/unten
Sollte sigma nicht gegeben sein:}
\usage{
binomialLattice(
  n,
  S0,
  K,
  r,
  v,
  sigma = NA,
  T,
  optionType,
  exerciseType,
  u = NA,
  d = NA
)
}
\description{
n = Anzahl Perioden
S0 = Heutige Aktienpreis
K = Strike
sigma = Volatilität
T = Periodizität: quartalsweise : T= 4, halbjährlich T = 2
optionType = Call/Put
exerciseType = Eu/Ami für Ami --> "American"
u = Preissprung nach oben
d = Preissprung nach unten
r = risikofreier Zins
v = erwarte Wachstumsrate der log-preise
Zeitschritte anwenden
Berechne Preissprünge oben/unten
Sollte sigma nicht gegeben sein:
}
