% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black_scholes.R
\name{black_scholes}
\alias{black_scholes}
\title{Black Sholes Formula}
\usage{
black_scholes(S, K, t, r, sigma, type = "call", short = FALSE, verbose = FALSE)
}
\arguments{
\item{S}{price of underlying}

\item{K}{strike price}

\item{t}{delta t (time)}

\item{r}{return}

\item{sigma}{volatility (standard deviation)}

\item{type}{either call or put [default = 'call']}

\item{short}{boolean (wheter option is shorted or not) [default = 'FALSE']}

\item{verbose}{boolean (print out information or not) [default = 'FALSE']}
}
\description{
Black Sholes Formula
}
