% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeks.R
\name{greeks}
\alias{greeks}
\title{S = Preis des Underlying
K = Strike
r = Zins
T = Fälligkeit
sig = Vola des Underlying
type = Call/Put}
\usage{
greeks(S, K, r, T, sig, type)
}
\description{
S = Preis des Underlying
K = Strike
r = Zins
T = Fälligkeit
sig = Vola des Underlying
type = Call/Put
}
