% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.binom.lattice.R
\name{plot.binom.lattice}
\alias{plot.binom.lattice}
\title{plot.binom.lattice}
\usage{
\method{plot}{binom.lattice}(assetlattice, n, optionValues = NA)
}
\arguments{
\item{optionValues}{}

\item{assetlattice:}{}

\item{n:}{maxmialer Zeitschritt}
}
\description{
nimmt Werte von der Funktion binomiallattice und plottet diese.
}
