% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_call_value.R
\name{put_call_value}
\alias{put_call_value}
\title{Value of a Eurpoean Put- or Call-Option}
\usage{
put_call_value(r, n, u, d, K, S0, type = "")
}
\arguments{
\item{r}{interest rate}

\item{n}{periods of the binomial lattice}

\item{u}{probability for upward movement in the binomial lattice}

\item{d}{probability for downward movement in the binomial lattice}

\item{K}{strike price}

\item{S0}{spot price}

\item{type}{either call or put}
}
\value{
value of put/call
}
\description{
Value of a Eurpoean Put- or Call-Option
}
