% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor.approx.R
\name{taylor.approx}
\alias{taylor.approx}
\title{Taylor Approximation}
\usage{
taylor.approx(S, S0, K, sig, r, T, type = "P")
}
\arguments{
\item{S:}{die verschiedenen S-Werte}

\item{S0:}{initialer S-Wert}

\item{K:}{Strike}

\item{sig:}{Volatilität}

\item{r:}{Zinssatz}

\item{type:}{Call oder Put}
}
\description{
Taylor Approximation
}
